+++ R/set_headers.R
@@ -87,6 +87,46 @@ as_new_data <- function(x, ..., values = NULL){
 # add header/footer row ----
 
 #' @export
+#' @title Add rows in body part
+#'
+#' @description Add rows in the flextable's body.
+#' It can be inserted at the top or the bottom.
+#' The function is column oriented, labels are specified for each columns, there
+#' can be more than a value - resulting in more than a new row.
+#'
+#' @param x a `flextable` object
+#' @param top should the row be inserted at the top or the bottom.
+#' @param ... a named list (names are data colnames) of strings
+#' specifying corresponding labels to add.
+#' @param values a list of name-value pairs of labels or values,
+#' names should be existing col_key values.
+#' If `values` is supplied argument `...` is ignored.
+#' @examples
+#' ft <- flextable(head(iris),
+#'   col_keys = c(
+#'     "Species", "Sepal.Length", "Petal.Length",
+#'     "Sepal.Width", "Petal.Width"
+#'   )
+#' )
+#'
+#' ft <- add_body(
+#'   x = ft, Sepal.Length = 1:5,
+#'   Sepal.Width = 1:5 * 2, Petal.Length = 1:5 * 3,
+#'   Petal.Width = 1:5 + 10, Species = "Blah", top = FALSE
+#' )
+#'
+#' ft <- theme_booktabs(ft)
+#' ft
+#' @seealso [flextable()], [add_header()], [add_footer()]
+add_body <- function(x, top = TRUE, ..., values = NULL){
+  if( !inherits(x, "flextable") )
+    stop("add_header supports only flextable objects.")
+  new_data <- as_new_data(x = x, ..., values = values)
+  x$body <- add_rows( x$body, new_data, first = top )
+  x
+}
+
+#' @export
 #' @title Add a rows of labels in header or footer part
 #'
 #' @description Add rows of labels in the flextable's header